Htilssu (#4)
* event listener

* full screen toggle

* socket setup
