+++ settings.gradle.kts
@@ -1,1 +1,9 @@
+pluginManagement {
+    plugins {
+        kotlin("jvm") version "2.0.0"
+    }
+}
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
+}
 rootProject.name = "Battle_Ship"
