+++ src/main/java/com/htilssu/BattleShip.java
@@ -1,17 +1,23 @@
 package com.htilssu;
 
 
-import com.htilssu.managers.EventManager;
-import com.htilssu.managers.ScreenManager;
+import com.htilssu.manager.ListenerManager;
+import com.htilssu.manager.ScreenManager;
 import com.htilssu.multiplayer.Client;
 import com.htilssu.multiplayer.Host;
-import com.htilssu.settings.GameSetting;
-import com.htilssu.utils.NetworkUtils;
+import com.htilssu.setting.GameSetting;
+import com.htilssu.util.AssetUtils;
+import com.htilssu.util.GameLogger;
 
 import javax.swing.*;
+import java.awt.*;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 
 
-public class BattleShip extends JFrame implements Runnable {
+public class BattleShip extends JFrame implements Runnable, KeyListener {
 
     /**
      * Số frame mỗi giây hiện tại
@@ -24,10 +30,11 @@ public class BattleShip extends JFrame implements Runnable {
     /**
      * Quản lý sự kiện
      */
-    private final EventManager eventManager = new EventManager();
+    private final ListenerManager listenerManager = new ListenerManager();
+    /**
+     * Quản lý host
+     */
     private final Host host = new Host();
-
-
     /**
      * Quản lý các màn hình trong game
      */
@@ -35,11 +42,12 @@ public class BattleShip extends JFrame implements Runnable {
     /**
      * Quản lý giao tiếp với server
      */
-    Client client = new Client();
+    Client client = new Client(this);
     /**
      * Biến đánh dấu có đang chạy thread render hay không
      */
     private boolean running;
+    private boolean isFullScreen = false;
 
 
     private BattleShip() {
@@ -48,12 +56,20 @@ public class BattleShip extends JFrame implements Runnable {
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         setLocationRelativeTo(null);
         add(screenManager.getCurrentScreen());
+        setIconImage(AssetUtils.loadAsset("/game_icon.png"));
+        addKeyListener(this);
+        setFocusable(true);
+        setResizable(false);
 
         pack();
-
         setUp();
     }
 
+    /**
+     * Lấy FPS hiện tại
+     *
+     * @return trả về số frame mỗi giây hiện tại
+     */
     public static int getCurrentFPS() {
         return currentFPS;
     }
@@ -67,7 +83,7 @@ public class BattleShip extends JFrame implements Runnable {
      * Hàm cài đặt các sự kiện
      */
     private void setUp() {
-//        eventManager.registerEvent(new PlayerShootEvent());
+
     }
 
     /**
@@ -77,7 +93,12 @@ public class BattleShip extends JFrame implements Runnable {
         running = true;
         thread.start();
         host.start();
-        NetworkUtils.find(5555);
+
+        client.scanHost();
+        client.getHostList().forEach(host -> {
+            client.connect(host, GameSetting.DEFAULT_PORT);
+        });
+
         setVisible(true);
     }
 
@@ -97,17 +118,16 @@ public class BattleShip extends JFrame implements Runnable {
         float nsPerFrame = 1e9f / GameSetting.FPS;
         int frame = 0;
         long lastTime = System.currentTimeMillis();
-        new Host();
-
 
         while (running) {
             long now = System.nanoTime();
+
             if (now - lastTickTime > nsPerTick) {
                 updateData();
                 checkEvent();
                 lastTickTime = now;
-
             }
+
             if (now - lastFrameTime > nsPerFrame) {
                 render();
                 frame++;
@@ -126,21 +146,77 @@ public class BattleShip extends JFrame implements Runnable {
      * Hàm check event
      */
     private void checkEvent() {
-        eventManager.checkEvents();
+
     }
 
+    /**
+     * Hàm render mỗi frame
+     */
     private void render() {
         setTitle("BattleShip - FPS: " + currentFPS);
         screenManager.getCurrentScreen().repaint();
     }
 
+    /**
+     * Hàm update dữ liệu mỗi tick
+     */
     private void updateData() {
 
     }
 
+    /**
+     * Thay đổi màn hình hiện tại
+     *
+     * @param screen màn hình cần chuyển đến
+     */
     public void changeScreen(int screen) {
         remove(screenManager.getCurrentScreen());
         repaint();
         screenManager.setCurrentScreen(screen);
+        add(screenManager.getCurrentScreen());
+        pack();
+        repaint();
     }
+
+    @Override
+    public void keyTyped(KeyEvent e) {
+
+    }
+
+    @Override
+    public void keyPressed(KeyEvent e) {
+        switch (e.getKeyCode()) {
+            case KeyEvent.VK_F11 -> {
+                toggleFullScreen();
+            }
+        }
+    }
+
+    @Override
+    public void keyReleased(KeyEvent e) {
+
+    }
+
+    private void updateScreenSize() {
+        screenManager.getCurrentScreen().setPreferredSize(new Dimension(GameSetting.WIDTH, GameSetting.HEIGHT));
+    }
+
+    private void toggleFullScreen() {
+        isFullScreen = !isFullScreen;
+        GraphicsDevice graphicsDevice = GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice();
+        dispose();
+        setUndecorated(isFullScreen);
+        if (isFullScreen) {
+            graphicsDevice.setFullScreenWindow(this);
+            setExtendedState(JFrame.MAXIMIZED_BOTH);
+            screenManager.getCurrentScreen().setPreferredSize(new Dimension(getWidth(), getHeight()));
+        } else {
+            graphicsDevice.setFullScreenWindow(null);
+            updateScreenSize();
+        }
+        pack();
+        setVisible(true);
+
+    }
+
 }
