+++ src/main/java/com/htilssu/entity/Player.java
@@ -1,0 +1,9 @@
+package com.htilssu.entity;
+
+public class Player {
+    String name;
+
+    public String getName() {
+        return name;
+    }
+}
