+++ src/main/java/com/htilssu/event/EventHandler.java
@@ -1,0 +1,10 @@
+package com.htilssu.event;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface EventHandler {
+
+
+}
