+++ src/main/java/com/htilssu/event/GameEvent.java
@@ -1,0 +1,17 @@
+package com.htilssu.event;
+
+public abstract class GameEvent {
+
+//    /**
+//     * Hàm thực hiện một chuỗi hành động khi {@link GameEvent#check()} trả về giá trị là {@code true}
+//     */
+//    public abstract void execute();
+//
+//    /**
+//     * Hàm check sự kiện khi nào thì sự kiện xảy ra
+//     *
+//     * @return trả về {@code true} nếu điều kiện thực hiện {@code Event} là hợp lệ, ngược lại
+//     * trả về {@code false}
+//     */
+//    public abstract boolean check();
+}
