+++ src/main/java/com/htilssu/listener/Listener.java
@@ -1,0 +1,8 @@
+package com.htilssu.listener;
+
+import com.htilssu.entity.Player;
+import com.htilssu.component.Position;
+
+public interface Listener {
+    void onPlayerShoot(Player player, Position position);
+}
