--- src/main/java/com/htilssu/managers/DifficultyManager.java
@@ -1,17 +1,0 @@
-package com.htilssu.managers;
-
-import java.util.Map;
-
-public class DifficultyManager {
-    private static final int NORMAL = 1;
-    private static final int HARD = 2;
-    public static final Map<Integer, Integer> difficultyMap = Map.of(
-            NORMAL, 10,
-            HARD, 15
-    );
-    public static int DIFFICULTY = DifficultyManager.NORMAL;
-
-    public static int getGameBoardSize() {
-        return difficultyMap.get(DIFFICULTY);
-    }
-}
