--- src/main/java/com/htilssu/managers/EventManager.java
@@ -1,24 +1,0 @@
-package com.htilssu.managers;
-
-import com.htilssu.events.GameEvent;
-
-import java.util.LinkedList;
-
-public class EventManager {
-    LinkedList<GameEvent> gameEvents = new LinkedList<>();
-
-    /**
-     * Kiểm tra các sự kiện trong game nếu 1 sự kiện xảy ra thì thực thi {@link GameEvent#execute()}
-     */
-    public void checkEvents() {
-        for (GameEvent event : gameEvents) {
-            if (event.check()) {
-                event.execute();
-            }
-        }
-    }
-
-    public void registerEvent(GameEvent event) {
-        gameEvents.add(event);
-    }
-}
