+++ src/main/java/com/htilssu/screen/SelfGrid.java
@@ -1,0 +1,15 @@
+package com.htilssu.screen;
+
+import com.htilssu.BattleShip;
+
+//Màn hình đặt thuyền
+public class SelfGrid {
+    private String name;
+    private BattleShip battleShip;
+
+    public SelfGrid(String name, BattleShip battleShip) {
+        super();
+        this.name = name;
+        this.battleShip = battleShip;
+    }
+}
