+++ src/main/java/com/htilssu/screen/SettingScreen.java
@@ -1,0 +1,34 @@
+package com.htilssu.screen;
+
+import javax.swing.*;
+
+public class SettingScreen extends JPanel {
+//    private Audio audioPlayer;
+//
+//    public SettingScreen(JPanel parentPanel, CardLayout cardLayout, Audio audioPlayer) {
+//        this.audioPlayer = audioPlayer;
+//        setLayout(null);
+//
+//        JLabel settingsLabel = new JLabel("Settings");
+//        settingsLabel.setFont(new Font("Arial", Font.BOLD, 24));
+//        settingsLabel.setBounds(100, 50, 200, 30);
+//        add(settingsLabel);
+//
+//        JLabel volumeLabel = new JLabel("Volume");
+//        volumeLabel.setBounds(100, 150, 100, 30);
+//        add(volumeLabel);
+//
+//        JSlider volumeSlider = new JSlider(JSlider.HORIZONTAL, -80, 6, 0);
+//        volumeSlider.setBounds(100, 200, 200, 50);
+//        volumeSlider.addChangeListener(e -> {
+//            int value = volumeSlider.getValue();
+//            audioPlayer.setVolume(value);
+//        });
+//        add(volumeSlider);
+//
+//        JButton backButton = new JButton("Back");
+//        backButton.setBounds(100, 300, 200, 50);
+//        backButton.addActionListener(e -> cardLayout.show(parentPanel, "MENU_SCREEN"));
+//        add(backButton);
+//    }
+}
