--- src/main/java/com/htilssu/screens/AttackGrid.java
@@ -1,5 +1,0 @@
-package com.htilssu.screens;
-
-//Màn hình tấn công thuyền
-public class AttackGrid {
-}
