--- src/main/java/com/htilssu/screens/SelfGrid.java
@@ -1,15 +1,0 @@
-package com.htilssu.screens;
-
-import com.htilssu.BattleShip;
-
-//Màn hình đặt thuyền
-public class SelfGrid {
-    private String name;
-    private BattleShip battleShip;
-
-    public SelfGrid(String name, BattleShip battleShip) {
-        super();
-        this.name = name;
-        this.battleShip = battleShip;
-    }
-}
